Don't force 'grab' style cursor
This works out better for the "drag handles" configuration.

Without this change, if you click on the draggable once,
but don't drag and don't "drop" it,
then the cursor (for the whole draggable element, not just the drag
handle) stays as "grab". (Because this style is leftover.)

Maybe this should be a configuration option... but for now, just
disabling it, because I'm handling the grab/grabbing cursor, on the
drag-handle, not on the whole draggable element.